+++ src/crypto/signing.h
@@ -27,6 +27,7 @@
 #ifndef ARK_CRYPTO_SIGNING_H
 #define ARK_CRYPTO_SIGNING_H
 
+#include <stdbool.h>
 #include <stddef.h>
 #include <stdint.h>
 
@@ -42,6 +43,7 @@ typedef struct signing_context_t {
     uint32_t    bip32Path[ADDRESS_MAX_BIP32_PATH];
     uint8_t     data[MAX_RAW_OPERATION];
     uint32_t    dataLength;
+    bool        isSchnorr;
 } SigningContext;
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -49,5 +51,10 @@ uint32_t signEcdsa(const cx_ecfp_private_key_t *privateKey,
                    const uint8_t *hash,
                    uint8_t *signature,
                    size_t signatureSize);
+
+////////////////////////////////////////////////////////////////////////////////
+uint32_t signSchnorr(const cx_ecfp_private_key_t *privateKey,
+                     const uint8_t *hash,
+                     uint8_t *signature);
 
 #endif  // #define ARK_CRYPTO_SIGNING_H